Update code examples for how to use the GoogleCloudResourceDetector. Delete the old deprecated resource detector in __init__.py, and replace it with the new resource detector from _detector.py. Add an entry point to setup.cfg so that the GCP resource detector works with auto instrumentation.
