+++ opentelemetry-resourcedetector-gcp/setup.cfg
@@ -39,3 +39,7 @@ where = src
 
 [options.extras_require]
 test =
+
+[options.entry_points]
+opentelemetry_resource_detector =
+    gcp_resource_detector = opentelemetry.resourcedetector.gcp_resource_detector:GoogleCloudResourceDetector
